rlbot/requirements.txt

gymnasium>=0.28.1
numpy>=1.24.0
torch>=2.0.0
stable-baselines3>=2.0.0


websockets>=10.4
nest-asyncio>=1.5.6
jsonschema>=4.17.3


Pillow>=9.4.0
pydantic>=2.0.0


rich>=13.3.1
python-dotenv>=1.0.0


tensorboard>=2.12.0
opencv-python>=4.7.0.72 


autopep8
black
